+++ validation/validation.js
@@ -1,10 +1,10 @@
-// TODO: Alerts in this file: ? if multiline comments, 1 if standalone_items, 5 else
+// TODO: Alerts in this file: ? if multiline comments, 1 if standalone_items, 6 else
 let a = "todo";
 //TODO: Alert
 let b = "todo";
 // TODO Alert
 let c = "todo";
-/* TODO: Alert */ 
+/* TODO: Alert */
 let d = "todo";
 /* TODO: Alert
 let alertThis = "TODO";